Some changes: Farness must be double not int!!

+++ src/test/java/it/bigdatalab/algorithm/BFSTest.java
@@ -490,7 +490,7 @@ public class BFSTest {
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("unCyclefarnessProvider")
-    void testFarness_UnCycle(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_UnCycle(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_undirected/32-cycle.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -510,13 +510,13 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double [] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("cyclefarnessProvider")
-    void testFarness_DiCycle(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiCycle(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32-cycle.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -536,13 +536,13 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("unWheelfarnessProvider")
-    void testFarness_UnWheel(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_UnWheel(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_undirected/32-wheel.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -562,13 +562,13 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("pathfarnessProvider")
-    void testFarness_DiPath(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiPath(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32-path.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -588,13 +588,13 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("tPathfarnessProvider")
-    void testAlgorithm_DiTPath(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testAlgorithm_DiTPath(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32t-path.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -614,13 +614,13 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("inStarfarnessProvider")
-    void testFarness_DiInStar(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiInStar(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32in-star.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -640,13 +640,13 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 
     @ParameterizedTest(name = "{index} => direction={0}, seeds={1}, nodes={2}, expected={3}")
     @MethodSource("outStarfarnessProvider")
-    void testFarness_DiOutStar(String direction, int[] seeds, int[] nodes, int[] expected) throws IOException, MinHash.SeedsException {
+    void testFarness_DiOutStar(String direction, int[] seeds, int[] nodes, double[] expected) throws IOException, MinHash.SeedsException {
         String path = new File("src/test/data/g_directed/32out-star.graph").getAbsolutePath();
         path = path.substring(0, path.lastIndexOf('.'));
         Parameter param = new Parameter.Builder()
@@ -667,7 +667,7 @@ public class BFSTest {
         RandomBFS algo = new RandomBFS(g, param.getNumSeeds(), param.getThreshold(), nodes, param.getNumThreads(), param.computeCentrality());
         Measure measure = algo.runAlgorithm();
 
-        int[] farness = measure.getFarness();
+        double[] farness = measure.getFarness();
         assertThat(expected).containsExactly(farness);
     }
 }
